## What

In the trash case of GET_TARGETS check that dependent resources like credentials exist before trying to access info about them.

Note that this only affects the trash case, because the regular case is already checking resources (eg ssh_credential) before trying to access info (eg credential_name, credential_uuid).

## Why

The idea was to speed up the GSA trashcan. The speed seems the same, but we might as well skip the extra queries.

With 1600 trash targets:
```
time o m m '<get_targets filter="rows=-1 sort=name" trash="1"/>' > /tmp/targ
main: 6.2s, 6.4s, 8.9s
  pr: 6.2s, 8.2s, 7,9s
```